ReadMe
A template for a ReadMe file to be updated as we work on the project
(though we can change what we want to include at any time). Currently
there is no code in the repository, so the only file that is mentioned
is some main which we probably will have.